stride

+++ src/models/nhits/nhits.py
@@ -144,7 +144,7 @@ class _NHITSBlock(nn.Module):
             if self.n_pool_kernel_size == 1:
                 self.pooling_layer = nn.Conv1d(1, 1, kernel_size=self.n_pool_kernel_size, stride=self.n_pool_kernel_size)
             else:
-                self.adjusted_stride = self.n_pool_kernel_size/2
+                self.adjusted_stride = int(self.n_pool_kernel_size/2)
                 self.pooling_layer = nn.Conv1d(1, 1, kernel_size=self.n_pool_kernel_size, stride=self.adjusted_stride)
 
         hidden_layers = []
